requirements.txt
aiofiles==25.1.0
anyio==4.12.1
argcomplete==3.6.3
async-lru==2.1.0
attrs==25.4.0
backports.zstd==1.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
boto3==1.42.49
botocore==1.42.49
brotli==1.2.0
cachetools==7.0.1
cattrs==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@88ade288431a46233f1556d1e141901b3ef0a36b
colorama==0.4.6
contourpy==1.3.3
coremltools==9.0
cross-web==0.4.1
cryptography==46.0.5
cycler==0.12.1
dacite==1.9.2
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
eventlet==0.40.4
exceptiongroup==1.3.1
fiftyone==1.13.0
fiftyone-brain==0.21.5
fiftyone_db==1.4.1
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
Flask-SocketIO==5.6.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
glob2==0.7
graphql-core==3.2.7
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanize==4.15.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
inflate64==1.0.4
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
labelImg==1.8.6
lap==0.5.12
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.16.0
networkx==3.6.1
numpy==2.3.5
onnx==1.20.1
onnx-ir==0.1.15
onnxruntime==1.24.1
onnxscript==0.6.0
onnxslim==0.1.85
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.0
paramiko==3.5.1
pillow==12.1.0
plotly==6.5.2
polars==1.38.1
polars-runtime-32==1.38.1
pprintpp==0.4.0
priority==2.0.0
protobuf==3.20.3
psutil==7.2.2
py7zr==1.1.0
pyaml==26.2.1
pybcj==1.0.7
pycparser==3.0
pycryptodomex==3.23.0
pydash==8.0.6
pymongo==4.9.2
PyNaCl==1.6.2
pyparsing==3.3.2
pyppmd==1.3.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2025.2
PyYAML==6.0.3
rarfile==4.2
regex==2026.1.15
requests==2.32.5
retrying==1.4.2
rtree==1.4.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.5.1
scipy==1.17.0
segmentation_models_pytorch==0.5.0
setuptools==81.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
sse-starlette==0.10.3
sseclient-py==1.9.0
starlette==0.52.1
strawberry-graphql==0.291.3
sympy==1.14.0
tabulate==0.9.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2026.2.15
timm==1.0.24
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer-slim==0.21.1
typing_extensions==4.15.0
tzlocal==5.3.1
ultralytics==8.4.14
ultralytics-thop==2.0.18
universal-analytics-python3==1.1.1
urllib3==2.6.3
voxel51-eta==0.15.3
wcwidth==0.6.0
Werkzeug==3.1.5
wrapt==2.1.1
wsproto==1.3.2
xmltodict==1.0.3
